axon/requirements.txt
flask
flask-cors
openai
chromadb
sentence-transformers
PyMuPDF
python-docx
pyttsx3
SpeechRecognition
duckduckgo-search
requests
beautifulsoup4
python-dotenv
tavily-python
markdown
weasyprint
pyperclip
mcp
